ref: restructure dataclasses and enums into model folder
